Fix some URL reversing breakages on Local Sites.
When generating the download links for files in the diff viewer, we were
calling reverse() instead of local_site_reverse(). This led to a failure
when used on a LocalSite. We now use the proper function.

Search results were also calling reverse(), instead of
local_site_reverse(), when attempting to take the user to the All Review
Requests page.

Reviewd at https://reviews.reviewboard.org/r/5847/