Fix displaying the review request draft banner after an edit.
A regression in one of the later RCs broke displaying the draft banner
after editing an existing published review request. The placeholder
banner in the DOM was being set prematurely as the active banner, and
never shown.

We now only add the banner up-front if it's visible, and we make sure to
show it when needed.

Fixes bug #3352
Reviewed at https://reviews.reviewboard.org/r/5848/