Keep the file index row heights consistent after loading.
When loading the files in the diff viewer, the height of the row in the
file index would fluctuate. The loaded row information was slightly
taller than when it just showed the spinner.

This was mostly due to the margin-bottom of the chunk dots. That margin
was unnecessary, even for wrapping, and has been removed.

There were also some margins for the graph icons that weren't really
needed, and a min-height on the row that didn't affect anything. I've
removed all this and standardized heights and offsets through variables.

Fixes bug #3361

Reviewed at https://reviews.reviewboard.org/r/5853/