Fix the extraction of commit hashes from Bitbucket webhook payloads.
Previously, the commit hash was read from the 'node' field of the JSON
payload, which only holds the first 12 characters of the git hash.

If we're trying to determine which review request ID corresponds to a
given webhook, but the request ID is not mentioned in the commit
comment, then we must search for an existing review request which
matches the commit hash in the webhook payload.

Since the 'get_review_request_id' function searches for an exact match
of the commit hash, we now use the 'raw_node' property, which contains
the full hash.

Testing Done:
Tested by pushing to a private Bitbucket git.

The commit did not mention the review request ID, but the request was
automatically closed after the webhook was received.

Reviewed at https://reviews.reviewboard.org/r/5851/