Escape '_' in all parts of a string when escaping for Markdown.
While our Markdown formatting through marked.js technically allows '_'
to be unescaped in the middle of words without problems, it breaks
down in practice. There are bugs in marked.js that can cause '_' in
different parts of a string to be turend into emphasis.

Until that's fixed, or we move to server-side Markdown rendering, we'll
need to disable the escaping of '_' within words in order to prevent
these problems.

Reviewed at https://reviews.reviewboard.org/r/5857/