Use "pre"-mode whitespace rules for <p> tags in Markdown.
Free-form text (<p> tags) are now set to not strip whitespace, causing
them to act just like a <pre> tag. This allows things like indentation
to remain (which Review Bot currently sets).

This is how it used to be, but that was changed in one of the betas to
work around some issues. However, that change was made to *all*
elements, and I don't believe the original issue was with paragraph
text.

Reviewed at https://reviews.reviewboard.org/r/5858/