Add functions for scanning accessible repositories on GitHub.
This change provides some functionality for listing all accessible
repositories on GitHub owned by the authenticated user, another user, or
an organization. This will eventually be used for the new repository
configuration UI.

This is based on Olessia's work as part of UCOSP in late 2013/early
2014. It updates her work for the GitHubClient changes recently made,
abstracts out API list pagination support, and fills in some of the
documentation.

Reviewed at https://reviews.reviewboard.org/r/5867/