Fix bumping Last Updated on a review request when publishing a review.
We made a change to improve the performance of publishing a review by
limiting what fields we were writing on a ReviewRequest. However, we
weren't including "last_updated" in the list of fields, which meant that
a review request with a new review wouldn't jump to the top of the
dashboard. Adding that field to the list fixes this.

Fixes bug #3374
Reviewed at https://reviews.reviewboard.org/r/5875/