Add an API resource for hosting services.
This introduces /api/hosting-services/, which provides basic information
on the hosting services supported by Review Board.

The resource does not expose all the information on a HostingService.
Right now, I'm keeping to the basic information, such as the name and
capabilities, until I determine what we're going to want to use for the
new repository configuration UI. This UI may impact how HostingService
itself structures data down the road, so it's not worth baking anything
related to that into the API yet.

This work is loosely based on Olessia's UCOSP work, though many
changes were made to better fit in with other resources, to take
advantage of what WebAPIResource provides, and to make use of the new
LocalDataQuerySet (which greatly reduces the amount of code).

Reviewed at https://reviews.reviewboard.org/r/5882/