Change the return type of get_hosting_services.
get_hosting_services was a generator returning tuples of (id, cls) for
each HostingService. This was to work around not having a way to access
the ID from a HostingService class.

This change fixes this by assigning the ID at registration time to the
class, allowing any consumer of the class to get the ID. This simplifies
get_hosting_services(), and is also needed for the new HostingService
API resource.

Reviewed at https://reviews.reviewboard.org/r/5883/