Add unit tests for HostingServiceAccountResource.
This resource never had any accompanying tests when it was first
written. As we made changes to the resource or any other backing
services, we had no way of verifying that the code worked through the
API. It turns out, it didn't, as POSTing ended up calling save() on the
wrong object.

This introduces a basic test suite for this resource.

Reviewed at https://reviews.reviewboard.org/r/5891/