Rename HostingServiceRepository to RemoteRepository and add new fields.
This standardizes on RemoteRepository as the name, which matches the
name of the upcoming API resource. It also adds some new fields to the
resource for storing the unique (per-service) ID of the repository and
the HostingService and account that the repository is linked to.

While in here, I noticed that the wrong field was being used to fetch
the repository's URL for GitHub. This updates to use `clone_url` instead
of `url`.

Reviewed at https://reviews.reviewboard.org/r/5896/