Change the API for get_remote_repositories().
get_remote_repositories()'s API was using plan IDs as a way of
identifying users vs. organizations, which led to some redundancy.
For example, "private" and "public" for GitHub meant the same thing
(users). GitHub in particular was assuming certain filter types for
repositories, which would be handy to specify from a caller's point of
view.

The function has now been changed to remove the plan parameter and to
instead have a dedicated "owner_type" parameter that specifies what
"owner" refers to. The value is dependent on the hosting service, and
HostingService subclasses are expected to be smart about the defaults.

It also now takes a "filter_type" parameter that allows for filtering
repositories. Possible values are dependent on the hosting service as
well. For instance, GitHub's allowed values consist of what their API
natively allows for their "type=" parameter in repository queries.

Reviewed at https://reviews.reviewboard.org/r/5903/