Add HostingService.get_remote_repository.
This allows fetching of an individual repository, with the given ID. The
IDs are specific to the HostingServices, and are provided when listing
repositories.

GitHub's remote repository IDs have been updated to provide both the
owner and repository name as the ID, instead of the numeric ID. This is
required in order to actually fetch repository information.

Reviewed at https://reviews.reviewboard.org/r/5913/