Changed Content-Disposition from inline to attachment for raw diff download.
When chrome is presented a file with Content-Disposition: inline, it shows the
file in the browser. This is fine for most things, but for the "Download Diff"
link, we'd prefer to offer to download the file.

Testing Done:
Manual testing with Chrome Version 35.0.1916.114, the affected browser. Also
tested Firefox 29.0. In Chrome, Download Diff now downloads instead of
displaying in the browser window. Added unit test.

Fixes bug 3384.

Reviewed at https://reviews.reviewboard.org/r/5893/