Add a profile field controlling whether users want to receive email.
Not everyone wants to get emails from Review Board. This change adds a flag to
the profile to control whether users will get anything directly addressed to
them.

This implementation was done on release-2.0.x because of the testing
infrastructure, but I also plan to backport this to release-1.7.x

Testing done:
* Ran unit tests
* Tested database evolution.
* Tested the account settings.

Reviewed at https://reviews.reviewboard.org/r/5907/