Fix a padding issue in the administration UI.
The top of every content area in the administration UI had excess
padding, due to the addition of padding-bottom on the 'navbar-container'
element. Non-admin pages were updated for this, but the admin pages were
not. This just removes the content area's padding-top on these pages.

Reviewed at https://reviews.reviewboard.org/r/5922/