Add a capability for whether a HostingServie can list remote repos.
This adds HostingService.supports_list_remote_repositories, which is
used for determining if get_remote_repositories() and
get_remote_repository() can be called. This defaults to false, but
GitHub enables it.

The API exposes this on HostingServiceResource, and
HostingServiceAccountResource uses it to determine whether to show a
'remote_repositories' link.

Reviewed at https://reviews.reviewboard.org/r/5923/