Fix SSL regressions with Subversion.
Subversion's SSL support badly broke with the subvertpy work done in
2.0. The new code flat-out didn't work with self-signed certificates,
and seemed to have trouble with verification all-together. This was true
of both PySVN and Subvertpy.

In PySVN's case, the SSL trust callback was never set. The base SVNTool
tried to set it on our Client class for PySVN, but it never got set on
the real Client class provided by PySVN.

Subvertpy was missing a bunch of provider classes that were necessary
for reading and writing the certificate verification files in the
.subversion directory.

In both cases, the new accept_certificate() functions were failing to
actually return the generated certificate. They were also not being
careful with the 'failures' key in the resulting certificate. PySVN
prematurely deleted it (causing us to fail when looking up the failures,
if needed), and Subvertpy was leaving it in.

This brings our support for SSL certificates back to a working state.

Fixes bug #3317
Reviewed at https://reviews.reviewboard.org/r/5936/