Allow for age-based filtering when reindexing for search.
This updates our SearchIndex to tell Haystack which field should be
considered to determine the age of a review request, in order to speed
up filtering.

Now, when running update_index with `--age=<hours>`, only review
requests that have been updated in <hours> will be updated or added to
the index.

Reviewed at https://reviews.reviewboard.org/r/5937/