Fix word wrapping in review request details fields.
Long strings in the review request details fields (on the right-hand side of
the fields) would increase the width of the entire table, eventually
overlapping the main fields in a very ugly way. This change adds a max-width to
the field spans using javascript to compute the correct width, and specifies a
word-wrap to force browsers to insert line breaks in the middle of words in the
case that it can't break at a natural place.

I've also made our layout recomputation method be throttled to 100ms, which
dramatically improves performance when resizing the window.

Testing done:
* Set a branch that was just one long unbroken string of characters and saw it
  wrap at the width of the details pane.
* Checked that a bunch of comma-separated bug numbers still wrapped at an ideal
  place in the text.
* Changed a label in a field to be longer and saw that the span max-width fit
  everything on the screen.
* Resized the window and saw that things worked correctly.

Fixes bug 3371.

Reviewed at https://reviews.reviewboard.org/r/5951/