Allow sending a test e-mail from the e-mail settings form.
When changing the server settings for how to send e-mail, it's nice to give
people the option to test it (so they don't have to create a fake review
request and jump through a bunch of hoops). This change adds a new check-box
that allows users to trigger such an email after the form is saved.

This doesn't cover 100% of the things that can go wrong with sending e-mails
(such as random SPF/DKIM failures or mail servers that don't like the
Sender/From distinction), but it should be good enough to verify that you typed
your SMTP credentials correctly.

This also organizes the e-mail settings into three sections.

Testing done:
Set my e-mail backend to console and saved the form with this check-box set.
Saw that it "sent" an e-mail from the correct Sender (noreply) to my user's
configured e-mail address.

Fixes bug 2073.

Reviewed at https://reviews.reviewboard.org/r/5952/