Switch the default SVN backend to PySVN and require a newer Subvertpy.
While Subvertpy is easier to build, it's less complete than PySVN, and
has been the source of several bugs. Some users who have tried switching
to PySVN from Subvertpy have had issues due to a version of Subvertpy
remaining on the system.

We now default to PySVN as the backend if it's installed. It's
better-tested and likely to be more compatible.

For Subvertpy support to work, a minimum version of 0.9.1 must be
installed, since older versions have caused errors we don't want to deal
with.

I also renamed 'imported_dependency' to 'has_svn_backend', which is a
bit more clear and less likely to result in conflicts.

Reviewed at https://reviews.reviewboard.org/r/5957/