Prevent crashes when unable to get/create a ReviewRequestVisit.
Under some unknown race condition, it's possible that looking up or
creating a ReviewRequestVisit will fail when viewing a review request.
This isn't the worst thing in the world, and is likely a temporary
problem (probably due to two requests for creating the same
ReviewRequestVisit at once).

Instead of crashing, we now log the failure and continue on.

Reviewed at https://reviews.reviewboard.org/r/5959/