Add a consistent function in the SVN backends for getting history logs.
PySVN and Subvertpy had very different ways of getting history logs from
a repository. Both had functions that accepted paths and revision
ranges, but they had different defaults for these parameters and
represented the results in different ways.

This adds a get_log() function for the SVN client backends that return
entries in consistent ways. It'll soon be used to centralize some of the
logic that is currently duplicated between the backends.

Reviewed at https://reviews.reviewboard.org/r/5978/