Add support for listing commits by branch ID.
Branches now contain a branch ID, which represents that specific branch.
In most cases, this would just be the name of the branch, but SVN uses a
relative path instead.

This branch ID can be passed to get_commits() to limit the commits to
those on the branch.

As of this change, this is not yet supported by the web API, but that's
coming up soon.

Reviewed at https://reviews.reviewboard.org/r/5986/