Fix filtering SVN commits by branch in the New Review Request page.
The New Review Request page wasn't properly filtering commits by branch
for Subversion repositories. It was assuming it could start at a
revision and pass that to the equivalent of 'svn log'. However, this
ended up picking up commits that existed on other branches as well.

The Repository Commits API for commits now accepts a branch= parameter,
which is the branch ID returned by the Repository Branches API. This is
passed through to get_commits(), which will then filter by that branch.

Through this, it's possible to properly limit the commits shown to those
on the branch. We'll also be able to make use of this later for other
repositories, if they suffer the same problem.

Fixes bug #3401
Reviewed at https://reviews.reviewboard.org/r/5987/