Adding a subversion repository accessed via https configured with basic authentication fails
Not every call to SVNTool.build_client passed the username and password, even
when they were available. This is a problem when trying to do things like
validate certificates.

Testing Done:
From the admin panel I tried to add a subversion repository accessed via https
configured with basic authentication fails with an "Authentication failed"
error. Tested with pysvn but it seems that the same may apply to subvertpy. The
username and password enter in the repository configuration are not passed when
validating the repository.

Reviewed at https://reviews.reviewboard.org/r/6034/