Add bug infobox support for GitHub.
This change adds an implementation of get_bug_info_uncached for GitHub
repositories. This is relatively straightforward (hurray for bug trackers which
give us sane results in json!). The one change I had to make was to make it so
Repository.bug_tracker_service returns an instance rather than a class, and
handles bug_tracker_use_hosting correctly.

Testing done:
Added the Review Bot repository to my server, and told it to use the hosting
service's bug tracker. Saw that links to the bugs worked, and that hovering
over bug IDs showed the infobox with the correct data.

Reviewed at https://reviews.reviewboard.org/r/6049/